[SEMVER_MINOR] Call end: overriding `freeswitch_linger` never worked. Replace with `cleanup_linger`.
This is only a minor semver because code that previously was capturing the documented events (`freeswitch_disconnect`, `freeswitch_linger`) won't see a difference.
The new code however deprecates capturing those events, and instead recommends using `cleanup_linger` and `cleanup_disconnect` as the proper ways to do this.
Since there is no universal way to capture the outcome of emitting an event (completion of the event handler might be asynchronous) we can't do `end()` or `exit()` if an event handler is present.