Use headers to tracks commands
Instead of using the parameters to match an `execute` command and its response, it appears we should be able to send an `Event-UUID` header and get it back in the `Application-UUID` header.

The value of the `Event-UUID` header is saved in the `app_uuid` variable (in `switch_ivr.c`) then retrieved (in `switch_core_session.c`) to populate the `Application-UUID` header.

Of course since this is a variable there might still be issues if the application manipulates it.